[hebo][打开GPS速度值，解决IQ相关编译不过问题]

+++ Project/DemoKit/SrcCode/UIWnd/SPORTCAM/UIFlow/UIFlowMovie/UIFlowMovieIcons.c
@@ -23,7 +23,7 @@ static CHAR    time_str[20] = {0};
 static CHAR    g_SpeedStr[20] = {0};
 static CHAR    g_DistanceStr[20] = {0};
 extern UINT32 uiGpsSpeed;
-extern UINT32 g_CurSpeed;
+extern FLOAT g_CurSpeed;
 #if _ADAS_FUNC_
 extern UINT32 g_uiAdasAlertDistance;
 #endif
